pa1/crawler/crawler_server/src/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.10
Werkzeug==2.2.2